feat: DialogContainer created
the current Dialogs are placed inside the DialogContainery
