+++ frontend/src/app/components/calendar/TimeSlotView.tsx
