+++ frontend/src/app/components/container/DogCardContainer.tsx
