+++ frontend/src/app/dogs/page.tsx
@@ -2,7 +2,7 @@
 import {Box, Container, Grid, Typography} from "@mui/material";
 import { getJWT} from "@/app/util/JWTDecoder";
 import Layout from "@/app/components/layout/Layout";
-import DogCardContainer from "@/app/components/cards/DogCardContainer";
+import DogCardContainer from "@/app/components/container/DogCardContainer";
 
 export default function Dogs() {
     const jwt = getJWT();
