+++ frontend/src/app/reservation/page.tsx
@@ -1,10 +1,10 @@
 'use client'
 import Layout from "@/app/components/layout/Layout";
-import FormContainer from "@/app/components/form/FormContainer";
+import FormContainer from "@/app/components/container/FormContainer";
 import GroomingForm from "@/app/components/form/GroomingForm";
 import {useState} from "react";
 import {Dog, Grooming} from "@/app/util/types";
-import Calendar from "@/app/components/Calendar";
+import Calendar from "@/app/components/calendar/Calendar";
 
 export default function Reservation() {
     const [dog, setDog] = useState<Dog|null>(null);
